requirements.txt
gradio>=4.0
yt-dlp
pydub
google-genai
youtube-transcript-api
